Amelioration on the render module
- Now there is no code duplication (before the params where all
rewriting, now it's directly pass to the object).
- the clear rect is now part of the loop, in a game there is a global
clean of the canvas, not a local for all the event.
- The sprite is center on the x and the y, before the x and the y where
at the corner.
- The x and the y don't have to be pass in the render method, the object
we want to have the render method as to have a x and y before add the
methods. If he don't have, the sprite will not be draw.
- the comment were confused so I clean them
- I better indent the code.